+++ test/Doc.hs
@@ -2,23 +2,65 @@
 
 module Main where
 
+import Data.Ord (comparing)
+import Data.List (intersect, sortBy)
 import Data.List.Split (splitOn)
-import BenchGraph (bgraph, defaultConfig, Config(..), ComparisonStyle(..))
+import BenchGraph
+       (defaultConfig, Config(..), ComparisonStyle(..),
+        SortField(..), graphCmp)
 
 main :: IO ()
 main = do
-    bgraph "test/results-doc.csv" "docs/mean-full" "mean" defaultConfig
     let classifier name =
             case splitOn "/" name of
                 grp : rest -> Just (grp, concat rest)
                 _      -> Nothing
-    bgraph "test/results-doc.csv" "docs/mean-grouped" "mean" defaultConfig
-        { classifyBenchmark = classifier }
-    bgraph "test/results-doc.csv" "docs/mean-grouped-delta" "mean" defaultConfig
+
+    graphCmp "test/results-doc.csv" "docs/full" defaultConfig
+    graphCmp "test/results-doc.csv" "docs/grouped"
+        defaultConfig
         { classifyBenchmark = classifier
-        , comparisonStyle = CompareDelta
+        , sortBenchFields = (`intersect` ["Mean"])
         }
-    bgraph "test/results-doc-multi.csv" "docs/mean-regression" "mean" defaultConfig
+    graphCmp "test/results-doc.csv" "docs/grouped-percent"
+        defaultConfig
         { classifyBenchmark = classifier
-        , comparisonStyle = CompareDelta
+        , comparisonStyle = ComparePercent
+        , sortBenchFields = (`intersect` ["Mean"])
+        }
+    graphCmp "test/results-doc.csv" "docs/grouped-delta"
+        defaultConfig
+        { classifyBenchmark = classifier
+        , comparisonStyle = CompareAbsoluteDiff
+        , sortBenchFields = (`intersect` ["Mean"])
+        }
+    graphCmp "test/results-doc.csv" "docs/grouped-percent-delta"
+        defaultConfig
+        { classifyBenchmark = classifier
+        , comparisonStyle = ComparePercentDiff
+        , sortBenchFields = (`intersect` ["Mean"])
+        }
+    graphCmp "test/results-doc-multi.csv" "docs/regression"
+        defaultConfig
+        { classifyBenchmark = classifier
+        , comparisonStyle = CompareAbsoluteDiff
+        , sortBenchFields = (`intersect` ["Mean"])
+        }
+    graphCmp "test/results-doc-multi.csv" "docs/regression-percent"
+        defaultConfig
+        { classifyBenchmark = classifier
+        , comparisonStyle = ComparePercentDiff
+        , sortBenchFields = (`intersect` ["Mean"])
+        }
+    graphCmp
+        "test/results-doc-multi.csv"
+        "docs/regression-percent-descending"
+        defaultConfig
+        { classifyBenchmark = classifier
+        , comparisonStyle = ComparePercentDiff
+        , sortBenchFields = (`intersect` ["Mean"])
+        , sortBenchmarks =
+              \f ->
+                  map fst $
+                  sortBy (comparing snd) $ f $ Name "streamly(2)(-base %)"
         }